IN PROGRESS - issue NXCM-1162: Get ITs working on grid https://issues.sonatype.org/browse/NXCM-1162
git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk@5053 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab

+++ nexus/nexus-test-harness/nexus-test-harness-launcher/src/main/java/org/sonatype/nexus/integrationtests/AbstractNexusIntegrationTest.java
@@ -432,9 +432,12 @@ public class AbstractNexusIntegrationTest
         catch ( Exception e )
         {
             log.fatal( e.getMessage(), e );
-            File testNexusLog = new File( nexusLogDir, getTestId() + "/nexus.log" );
-            testNexusLog.getParentFile().mkdirs();
-            FileUtils.copyFile( nexusLog, testNexusLog );
+            if ( nexusLog.exists() )
+            {
+                File testNexusLog = new File( nexusLogDir, getTestId() + "/nexus.log" );
+                testNexusLog.getParentFile().mkdirs();
+                FileUtils.copyFile( nexusLog, testNexusLog );
+            }
             throw e;
         }
     }
